Update #38
CustomRecordParser has been deprecated, use the HermitCoreConfig class
for this now(the record parser didn't work to well, but will be readded
in the future to add the new records to the sounds.json for the mod, so
the sounds a re loaded in the songs are loaded in the game when playing
the records in a jukebox.
Will begin work on a lang parser that will grab the en_US.lang from the
config folder so i can have java add the language definition for the new
records.